Bug 1825705 - Move PerfData to a new component
